intptr_t

+++ elements/wifi/athdescencap.cc
@@ -91,7 +91,7 @@ AthdescEncap_write_param(const String &in_s, Element *e, void *vparam,
 {
   AthdescEncap *f = (AthdescEncap *)e;
   String s = cp_uncomment(in_s);
-  switch((int)vparam) {
+  switch((intptr_t)vparam) {
   case H_DEBUG: {    //debug
     bool debug;
     if (!cp_bool(s, &debug))
